Get it to build in MSVC.

+++ src/Models.h
@@ -21,7 +21,7 @@ struct PositionsAndElements {
 PositionsAndElements icosahedron();
 PositionsAndElements icosphere(float radius, int refinements);
 
-extern const float ICOSAHEDRON_VERTICES[12][3];
+extern const double ICOSAHEDRON_VERTICES[12][3];
 extern const unsigned int ICOSAHEDRON_VERTEX_COUNT;
 extern const unsigned int ICOSAHEDRON_ELEMS[60];
 extern const unsigned int ICOSAHEDRON_ELEM_COUNT;
